+++ BubbleBobble/Buffer.cpp
@@ -22,7 +22,7 @@ bool Buffer::LoadFile()
 	if (mpData == nullptr)
 	{
 		std::ifstream ifile{ "../Data/Amiga/" + mFilename, std::ios::binary | std::ios::ate };
-		mSize = size_t(ifile.tellg());
+		mSize = int(ifile.tellg());
 		ifile.seekg(0);
 		mpData = new char[mSize];
 		ifile.read(mpData, mSize);
