+++ BubbleBobble/BufferBubble.cpp
@@ -5,28 +5,28 @@
 
 using namespace ieg;
 
-const size_t BufferBubble::mIntroPaletteOffset{ 0x3af8 };
-const size_t BufferBubble::mPaletteColorCount{ 16 };
-const size_t BufferBubble::mPaletteOffset{ 0x7b6 };
-const size_t BufferBubble::mLevelColorCount{ 4 };
-const size_t BufferBubble::mLevelColorsOffset{ 0x275C };
-const size_t BufferBubble::mLevelColorsPaletteOffset{ 11 };
-const size_t BufferBubble::mLevelBigBlockOffsets{ 0x2418 };
+const int BufferBubble::mIntroPaletteOffset{ 0x3af8 };
+const int BufferBubble::mPaletteColorCount{ 16 };
+const int BufferBubble::mPaletteOffset{ 0x7b6 };
+const int BufferBubble::mLevelColorCount{ 4 };
+const int BufferBubble::mLevelColorsOffset{ 0x275C };
+const int BufferBubble::mLevelColorsPaletteOffset{ 11 };
+const int BufferBubble::mLevelBigBlockOffsets{ 0x2418 };
 
 BufferBubble::BufferBubble(const std::string& filename)
 	: Buffer(filename)
 {
 }
 
-size_t BufferBubble::GetPaletteColorCount()
+int BufferBubble::GetPaletteColorCount()
 {
 	return mPaletteColorCount;
 }
 
-void BufferBubble::GetPalette(ColorRGBA8* pPalette, const size_t offset)
+void BufferBubble::GetPalette(ColorRGBA8* pPalette, const int offset)
 {
 	ColorXRGB4* pColor{};
-	for (size_t i{ 0 }; i < mPaletteColorCount; ++i)
+	for (int i{ 0 }; i < mPaletteColorCount; ++i)
 	{
 		pColor = (ColorXRGB4*)(mpData + offset + i * sizeof(ColorXRGB4));
 		pColor->ConvertRGBA8(pPalette + i);
@@ -38,20 +38,20 @@ void BufferBubble::GetIntroColors(ColorRGBA8* pPalette)
 	GetPalette(pPalette, mIntroPaletteOffset);
 }
 
-void BufferBubble::GetLevelColors(ColorRGBA8* pPalette, size_t level)
+void BufferBubble::GetLevelColors(ColorRGBA8* pPalette, int level)
 {
 	GetPalette(pPalette, mPaletteOffset);
 	ColorXRGB4* pColor{};
-	for (size_t i{ 0 }; i < mLevelColorCount; ++i)
+	for (int i{ 0 }; i < mLevelColorCount; ++i)
 	{
 		pColor = (ColorXRGB4*)(mpData + mLevelColorsOffset + (level * mLevelColorCount + i) * sizeof(ColorXRGB4));
 		pColor->ConvertRGBA8(pPalette + mLevelColorsPaletteOffset + i);
 	}
 }
 
-size_t BufferBubble::GetLevelBigBlockOffset(size_t level)
+int BufferBubble::GetLevelBigBlockOffset(int level)
 {
 	unsigned int offset{ *(((unsigned int*)(mpData + mLevelBigBlockOffsets)) + level) };
 	offset = _byteswap_ulong(offset);
-	return size_t(offset);
+	return int(offset);
 }
